refactor: consolidate datasource modules and migrations
